+++ grunt/options/jscs.js
@@ -1,0 +1,15 @@
+/**
+ * Configure JSCS task
+ */
+'use strict';
+
+var config = require('../config');
+
+module.exports = {
+  js: {
+    src: config.js.hintFiles,
+    options: {
+      config: '.jscsrc'
+    }
+  }
+};