+++ src/js/plugins/console.js
@@ -7,12 +7,14 @@
   var method;
   var methodsToEnhance = ['log', 'warn'];
   var noop = function () {};
+
   var methods = [
     'assert', 'clear', 'count', 'debug', 'dir', 'dirxml', 'error',
     'exception', 'group', 'groupCollapsed', 'groupEnd', 'info', 'log',
     'markTimeline', 'profile', 'profileEnd', 'table', 'time', 'timeEnd',
     'timeline', 'timelineEnd', 'timeStamp', 'trace', 'warn'
   ];
+
   var length = methods.length;
   var console = (window.console = window.console || {});