+++ src/js/plugins/tracking.js
@@ -23,7 +23,9 @@ export default function () {
   var document = window.document;
 
   /* jshint ignore:start */
+  // jscs:disable
   (function(b,o,i,l,e,r){b.GoogleAnalyticsObject=l;b[l]||(b[l]=function(){(b[l].q=b[l].q||[]).push(arguments)});b[l].l=+new Date;e=o.createElement(i);r=o.getElementsByTagName(i)[0];e.src='//www.google-analytics.com/analytics.js';r.parentNode.insertBefore(e,r)}(window,document,'script','ga'));
+  // jscs:enable
   /* jshint ignore:end */
 
   // Create tracker object
@@ -50,7 +52,7 @@ export default function () {
       url = $target.attr('href');
       value = parseInt($target.attr('data-ga-value'), 10) || undefined;
 
-      if (ga !== undefined && category !== undefined && action !== undefined ) {
+      if (ga !== undefined && category !== undefined && action !== undefined) {
 
         // Register the event
         ga('send', 'event', category, action, label, value);